pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "badbunny-monitor"
version = "0.1.0"
description = "Telegram bot + TicketSwap monitor for Bad Bunny Madrid listings"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "BadBunny Maintainers" }
]
dependencies = [
  "python-telegram-bot>=21.0,<22",
  "httpx>=0.27,<0.29",
  "beautifulsoup4>=4.12,<5",
  "python-dotenv>=1.0,<2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<0.25"
]

[project.scripts]
badbunny-monitor = "badbunny_monitor.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
